--- pkg/session/pipelineddmltest/BUILD.bazel
@@ -1,24 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_test")
-
-go_test(
-    name = "pipelineddmltest_test",
-    timeout = "short",
-    srcs = [
-        "main_test.go",
-        "pipelineddml_test.go",
-    ],
-    flaky = True,
-    shard_count = 13,
-    deps = [
-        "//pkg/config",
-        "//pkg/kv",
-        "//pkg/sessionctx/binloginfo",
-        "//pkg/testkit",
-        "//pkg/testkit/testmain",
-        "//pkg/testkit/testsetup",
-        "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_stretchr_testify//require",
-        "@com_github_tikv_client_go_v2//tikv",
-        "@org_uber_go_goleak//:goleak",
-    ],
-)
