+++ tests/realtikvtest/txntest/BUILD.bazel
@@ -6,16 +6,19 @@ go_test(
     srcs = [
         "isolation_test.go",
         "main_test.go",
+        "pipelineddml_test.go",
         "txn_state_test.go",
         "txn_test.go",
     ],
     flaky = True,
     race = "on",
     deps = [
+        "//pkg/config",
         "//pkg/expression",
         "//pkg/kv",
         "//pkg/parser",
         "//pkg/session/txninfo",
+        "//pkg/sessionctx/binloginfo",
         "//pkg/store/driver",
         "//pkg/testkit",
         "//pkg/util",
